Separate plugin and Matrix from StarPU
